# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "API checker",
    "version": "1.0.0"
  },
  "host": "localhost:3001",
  "schemes": [
    "http"
  ],
  "paths": {
    "/customers": {
      "post": {
        "summary": "Create a new customer",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "name",
            "in": "rawData",
            "description": "Customer name",
            "required": true,
            "type": "string"
          },
          {
            "name": "email",
            "in": "rawData",
            "description": "Customer email",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "Customer created successfully"
          },
          "400": {
            "description": "Invalid input data"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "get": {
        "summary": "Get a list of all customers",
        "responses": {
          "200": {
            "description": "List of customers",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "put": {
        "summary": "Update an existing customer",
        "parameters": [
          {
            "name": "id",
            "in": "string",
            "description": "Customer ID",
            "required": true,
            "type": "string"
          },
          {
            "name": "name",
            "in": "formData",
            "description": "Customer name",
            "required": true,
            "type": "string"
          },
          {
            "name": "email",
            "in": "formData",
            "description": "Customer email",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Customer updated successfully"
          },
          "400": {
            "description": "Invalid input data"
          },
          "404": {
            "description": "Customer not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "summary": "Delete an existing customer",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Customer ID",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Customer deleted successfully"
          },
          "400": {
            "description": "Invalid input data"
          },
          "404": {
            "description": "Customer not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }}}}}